Assume that rules with output interfaces not part of the routing table never match
This has an effect if a routing table was specified only. The current implementation
keeps the interfaces if they are not part of the routing table. Due to the over
approximation, this assumes that traffic is allowed while it is not.

This modifieds the generated code which is a bad practice. However, I could not make
isabelle generate code at all.